refactor(cloud-workers): keep inventory helpers internal
