+++ src/gateway/worker-environments/workspace-accepted-sync.ts
@@ -15,7 +15,7 @@ import {
   serializeWorkerWorkspaceManifest,
   type WorkerWorkspaceManifest,
 } from "./workspace-manifest.js";
-import { changedPaths, manifestNodes } from "./workspace-reconcile-plan.js";
+import { changedPaths, manifestNodes } from "./workspace-reconcile.js";
 import {
   captureRemoteWorkspaceManifest,
   WORKER_WORKSPACE_RSYNC_DESTINATION,
