+++ src/gateway/worker-environments/workspace-inventory-limits.ts
@@ -1,0 +1,5 @@
+export const MAX_WORKSPACE_INVENTORY_ENTRIES = 250_000;
+export const MAX_WORKSPACE_GIT_CANDIDATES = MAX_WORKSPACE_INVENTORY_ENTRIES * 4;
+export const MAX_WORKSPACE_INVENTORY_PATH_BYTES = 64 * 1024 * 1024;
+export const MAX_WORKSPACE_MANIFEST_BYTES = 64 * 1024 * 1024;
+export const MAX_WORKSPACE_INVENTORY_TOTAL_BYTES = 4 * 1024 * 1024 * 1024;
