+++ src/gateway/worker-environments/workspace-reconcile-apply.ts
@@ -16,8 +16,10 @@ import {
 import {
   applyWorkspaceDirectoryChanges,
   assertActualWorkspaceManifest,
+  changedPaths,
   ConcurrentWorkspacePathError,
   hasReplacedBaseEntryAncestor,
+  manifestNodes,
   preflightWorkspaceApply,
   readActualWorkspaceManifest,
   retainedConflictPaths,
@@ -29,7 +31,6 @@ import {
   reconciliationEntries,
 } from "./workspace-reconcile-derived-paths.js";
 import { entryMatches } from "./workspace-reconcile-fs.js";
-import { changedPaths, manifestNodes } from "./workspace-reconcile-plan.js";
 import {
   applyWorkspacePatch,
   createWorkspacePatch,
